Create the basic project structure and mock out cmds. The basic usage is:
./ghub [verb] [noun] [flags and modifiers]

The project structure reflects that. Main contains the verbs, and
the nouns are contained in their own packages.

So for users, for example, the command would be:

./ghub create user [username]

The 'create' command is in main(), and the 'user' command is in the
'user' package.

Each package is then organized into a [noun]_cli.go file and a
[noun]_actions.go file. The cli.go files contain just the cli
call structure, while the _actions.go contain the functions that
the cli calls.

Still learning how to do all this, so it will likely change, but
seems clean to me.